agent-openai-agents-sdk-multiagent/pyproject.toml
[project]
name = "agent-server"
version = "0.1.0"
description = "Multi-agent orchestrator that queries other agents, Genie spaces, and serving endpoints"
readme = "README.md"
authors = [
    { name = "Agent Developer", email = "developer@example.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
    "databricks-openai>=0.13.0",
    "databricks-agents>=1.9.3",
    "mlflow>=3.10.0",
    "openai-agents>=0.4.1",
    "python-dotenv>=1.2.1",
    "a2a-sdk>=0.2.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[dependency-groups]
dev = [
    "hatchling>=1.28.0",
    "pytest>=9.0.2",
]

setup = [
    "ruamel.yaml>=0.18.0",
]

[tool.uv]
default-groups = ["dev", "setup"]


[project.scripts]
quickstart = "scripts.quickstart:main"
start-app = "scripts.start_app:main"
start-server = "agent_server.start_server:main"
agent-evaluate = "agent_server.evaluate_agent:evaluate"
discover-tools = "scripts.discover_tools:main"
preflight = "scripts.preflight:main"

